#10854 remove htmlpage table references when doing pp and integrity checker
